+++ tests/test_obj_address.py
@@ -4,7 +4,7 @@ import ecdsa
 import hashlib
 
 from aergo.herapy.obj.address import Address
-from aergo.herapy.utils.encoding import decode_address
+from aergo.herapy.utils.encoding import encode_address, decode_address
 from aergo.herapy.utils.converter import convert_public_key_to_bytes
 
 
@@ -84,3 +84,32 @@ def test_success():
     assert str(addr4) == str(addr5)
     assert bytes(addr4) == bytes(addr5)
     assert addr4.public_key.point == addr5.public_key.point
+
+
+def test_encode_empty():
+    addr = Address(None, empty=True)
+    assert str(addr) == ''
+
+    addr = Address(None, empty=True)
+    addr.value = bytes([])
+    assert str(addr) == ''
+
+
+def test_govname():
+    addr = Address(None, empty=True)
+    name = 'abcdetest'
+    addr.value = name.encode()
+    assert len(str(addr)) == len(name)
+
+    addr = Address(None, empty=True)
+    baddr = bytes([0x61, 0x65, 0x72, 0x67, 0x6f, 0x2e, 0x65, 0x6e, 0x74, 0x65, 0x72, 0x70, 0x72, 0x69, 0x73, 0x65])
+    addr.value = baddr
+    assert str(addr) == 'aergo.enterprise'
+    assert str(addr) == Address.encode(baddr)
+    assert bytes(addr) == Address.decode('aergo.enterprise')
+    assert baddr == Address.decode('aergo.enterprise')
+
+    addr = Address(None, empty=True)
+    addr.value = 'aergo.system'
+    assert str(addr) == 'aergo.system'
+    assert bytes(addr) == Address.decode('aergo.system')